+++ web_helper/frm_url_fix_set.cs
@@ -126,7 +126,7 @@ namespace web_helper
         private void txt_cell_TextChanged(object sender, EventArgs e)
         {
             if (string.IsNullOrEmpty(this.lb_column.Text) || string.IsNullOrEmpty(this.lb_row.Text)||this.dgv_result.Rows.Count==0) return;
-            this.dgv_result.Rows[Int16.Parse(this.lb_row.Text)].Cells[Int16.Parse(this.lb_column.Text)].Value = this.txt_cell.Text;
+            this.dgv_result.Rows[Int32.Parse(this.lb_row.Text)].Cells[Int32.Parse(this.lb_column.Text)].Value = this.txt_cell.Text;
         }
